Add DispenseableResource (+testcase)
Its based on the idea of RedbeanPHP (ORM Mapper) and supports
rapid prototyping. Basic example:

$disRes->init();
$disRes->foo = "bar";
$disRes->save();

Now $disRes produces two triples in the store, one said it is a
RDF resource and the other said that this resource has a property
called foo.

The URI for the resource is generated via unique hash. And the foo
property uses that URI as prefix.