Name test files by their type (integration or unit tests); remove cache unit tests
Naming was neccessary because we had mixed group of files, some
with IntegrationTest suffix, some with Test. We need this suffix,
because some classes have both integration and unit tests so
we cant use for instance FoobarTest 2 times, because an
"class FoobarTest was already declared"-error occours.

Remove unit tests for Cache adapter, because they were the same
as the integration tests.