Fix testDeleteMultipleStatements in AbstractTriplePatternStoreTest
Using as it is but expect an exception. Reason for that
is, that the query function rerouts to the
deleteMatchingStatements function which wants only a
Statement and not a set of Statement.

Add simple triple recognition test too, because the
testDeleteMultipleStatements was testing something with
triples. And because we have no other tests with triples
which cover that, i added it.