Port cache related classes to factory approach #17 #18 #20
From now on, you use CacheFactoryImpl to create cache
backends. Method createCache needs an configuration array
which now contains a class string. It represents the
class of the cache backend you want to instantiate.

CacheFactoryImpl tries to find the class and if it
exists, it will create it with given $config array.

Furthermore:
- Removed remaining PHPArrayCache stuff.
- Adapted affected query cache tests
- Adapted test-config.yml.dist and set classes for file-
  and memcached
- Also use fileCacheConfig for standardCache