Memcached has to check if the service is available #8
In the constructor is now a small code portion, which
set and get something in the cache. if NOT the same value
that was set, was returned, an exception will be thrown.

In the CacheMemcacheDTest.php setUp method is now a
try-catch around the fixture init to catch that exception
and mark the tests skipped instead of just break up with
a nasty exception.