TestCase: add assertIteratorContent to compare Iterator instances
The problem is, that if you compare two instances, which
both implement \Iterator, you can simply sort them, but
otherwise PHPUnit sometimes say, they are not the same,
even if they contain the same elements, but not in the
same order.

This function helps with that. It first checks, if
$expected contains the same elements as $actual.
Afterwards it counts the elements. If there are differences
they will be shown using assert-functions from PHPUnit.