Check HTTP store, what you can do and what not (skip tests, if neccessary) #30
HttpTest now checks for each test, if the target store supports
graph creation/deletion and triple querying/creation/deletion. If
not, the test will be marked as skipped.

The according check-method called getRights is very basic and
you could end up with empty graphs, just created to test something.
Unless, there is no better way how to check all the stuff, the
method will not be part of the constructor.