Remove existing code
Sorry Viktor. Reason for that is that it is too tight to be
extended easily. We want the REST api to be more like a
middleware which only computes given requests and returns a
response. To do so, it must be PSR-7 compatible, to provide
the most flexible approach.

Response and request implementation must have to implement
ResponseInterface and RequestInterface from PSR-7, so you
can use Slim, Symfony ... as the underlying middleware/framework.