Get rid of Streamer (for now, maybe come back later)
According issue: https://github.com/fzaninotto/Streamer/issues/5
Problem was, that Stream closes the stream on __destruct(). so
if we want to use it inside a function, it closes the stream after
the function "finished".

I port the code to basic file operations like fopen, fwrite.
Because the according functions want a stream, via resource or
string, we do not close it after we are finished. Even if there
is nobody doing it later on, PHP will solve that after the script
finished.

API-change: extend $outputStream, it can be a string or a resource.
So no backward compatibility break.