Saft/Rdf/LiteralImpl: change type of $datatype parameter in constructor
Inside the __construct method we already check, if that instance
is of type NamedNode, otherwise an exception will be thrown.

Using that old version will let you run into problems using
dependency injection, because the loader expects an instance of
Node.

Because of that, a change from Node => NamedNode was done.