broken BC warning; marked classes deprecated, changed according code and tests
* added constructQuery option to AbstractTriplePatternStore's
query function

* added transformStatementSetResultToStatementIterator to
Saft/Skeleton/Rest/Hub.php

* Broken backward compatibility: BasicTriplePatternStore
will return StatementSetResultImpl instead of
ArrayStatementIteratorImpl from now on.

## Deprecation warnings for NQuadsSerializerImpl and
SerializerFactoryImpl

### NQuadsSerializerImpl

We marked Saft/Data/NQuadsSerializerImpl.php as deprecated, because it
needs more refinement, but with EasyRDF a good alternative is already
available. We leave the current version available, but we expect to
remove it in one of the later releases. It wasn't part of the API
before, but we wanted to give you guys a bit more heads-up.

### SerializerFactoryImpl

Our SerializerFactoryImpl will **not** provide an instance of this class
anymore! Furthermore, it will be marked deprecated too, because without
the n-quads serializer it has no other purpose. Its function
*createSerializerFor* will throw an exception when used.

### Saft\Skeleton\Data\SerializerFactory

Class SerializerFactory adapted to not use SerializerFactoryImpl class
anymore. It will only use SerializerFactoryEasyRdf from now on.

## Broke backward compatibility in BasicTriplePatternStore

To fix an invalid state in **Store/BasicTriplePatternStore.php**, we
changed its return value for some functions from
ArrayStatementIteratorImpl.php to StatementSetResultImpl. This change
reflects the fact, that the query function only returns an instance of
Sparql\Result\Result.php.